requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiogram==3.4.1
python-dotenv==1.0.0
openai==1.12.0
chromadb==0.4.22
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
apscheduler==3.10.4
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2


